requirements.txt
selenium==4.15.2
requests==2.31.0
fake-useragent==1.4.0
beautifulsoup4==4.12.3
2captcha-python==1.2.1
python-dotenv==0.21.0
cryptography==42.0.5
psycopg2-binary
